150686 by jwiele on 2021/06/25 (pair: jshimkus) VDOUP-6 Remove PRI* print macros.
	The PRI* print macros can be replaced with standard format
	directives with appropriate casts for platform defined integer
	types. For example, uint64_t may be defined as "unsigned long"
	or "unsigned long long" but the latter is guaranteed to always
	be at least 64 bits, so a uint64_t quantity may be printed
	safely by casting to "unsigned long long and formatting with
	the "%llu" directive. See change 150333 for more details.

	.../allocatingVIO.h
	.../allocatingVIO.c
	.../blockAllocator.c
	.../blockMap.c
	.../blockMapFormat.c
	.../blockMapTree.c
	.../dataVIO.h
	.../dataVIO.c
	.../flush.c
	.../hashZone.c
	.../logicalZone.c
	.../packer.c
	.../physicalZone.c
	.../recoveryJournal.c
	.../recoveryJournalBlock.c
	.../recoveryUtils.c
	.../refCounts.c
	.../referenceCountRebuild.c
	.../slab.c
	.../slabDepot.c
	.../slabDepotFormat.c
	.../slabJournal.c
	.../slabScrubber.c
	.../vdo.c
	.../vdoComponent.c
	.../vdoComponentStates.c
	.../vdoPageCache.c
	.../vdoRecovery.c
	.../vdoResize.c
	.../vdoResizeLogical.c
	.../vio.c
	.../vioPool.c
	.../vioWrite.c
	.../dmvdo.c
	.../vdo/messageStats.c
	.../vdo/poolSysfsStats.c
	Removed use of PRI* print macros, replacing them with standard
	  print format characters, adding casts where necessary for
	  platform dependent types.

+++ vdo/allocatingVIO.c
@@ -16,7 +16,7 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA
  * 02110-1301, USA. 
  *
- * $Id: //eng/linux-vdo/src/c++/vdo/base/allocatingVIO.c#38 $
+ * $Id: //eng/linux-vdo/src/c++/vdo/base/allocatingVIO.c#39 $
  */
 
 #include "allocatingVIO.h"
@@ -64,7 +64,7 @@ static int attempt_pbn_write_lock(struct allocating_vio *allocating_vio)
 		// This block is already locked, which should be impossible.
 		return log_error_strerror(VDO_LOCK_ERROR,
 					  "Newly allocated block %llu was spuriously locked (holder_count=%u)",
-					  allocating_vio->allocation,
+					  (unsigned long long) allocating_vio->allocation,
 					  lock->holder_count);
 	}
