+++ vdo/dataVIO.h
@@ -16,7 +16,7 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA
  * 02110-1301, USA. 
  *
- * $Id: //eng/linux-vdo/src/c++/vdo/base/dataVIO.h#85 $
+ * $Id: //eng/linux-vdo/src/c++/vdo/base/dataVIO.h#86 $
  */
 
 #ifndef DATA_VIO_H
@@ -634,7 +634,7 @@ static inline void assert_data_vio_in_hash_zone(struct data_vio *data_vio)
 	// identifier.
 	ASSERT_LOG_ONLY((expected == thread_id),
 			"data_vio for logical block %llu on thread %u, should be on hash zone thread %u",
-			data_vio->logical.lbn,
+			(unsigned long long) data_vio->logical.lbn,
 			thread_id,
 			expected);
 }
@@ -681,7 +681,7 @@ static inline void assert_data_vio_in_logical_zone(struct data_vio *data_vio)
 	thread_id_t thread_id = vdo_get_callback_thread_id();
 	ASSERT_LOG_ONLY((expected == thread_id),
 			"data_vio for logical block %llu on thread %u, should be on thread %u",
-			data_vio->logical.lbn,
+			(unsigned long long) data_vio->logical.lbn,
 			thread_id,
 			expected);
 }
@@ -769,7 +769,7 @@ static inline void assert_data_vio_in_duplicate_zone(struct data_vio *data_vio)
 	thread_id_t thread_id = vdo_get_callback_thread_id();
 	ASSERT_LOG_ONLY((expected == thread_id),
 			"data_vio for duplicate physical block %llu on thread %u, should be on thread %u",
-			data_vio->duplicate.pbn,
+			(unsigned long long) data_vio->duplicate.pbn,
 			thread_id,
 			expected);
 }
@@ -816,7 +816,7 @@ static inline void assert_data_vio_in_mapped_zone(struct data_vio *data_vio)
 	thread_id_t thread_id = vdo_get_callback_thread_id();
 	ASSERT_LOG_ONLY((expected == thread_id),
 			"data_vio for mapped physical block %llu on thread %u, should be on thread %u",
-			data_vio->mapped.pbn,
+			(unsigned long long) data_vio->mapped.pbn,
 			thread_id,
 			expected);
 }
@@ -849,7 +849,7 @@ static inline void assert_data_vio_in_new_mapped_zone(struct data_vio *data_vio)
 	thread_id_t thread_id = vdo_get_callback_thread_id();
 	ASSERT_LOG_ONLY((expected == thread_id),
 			"data_vio for new_mapped physical block %llu on thread %u, should be on thread %u",
-			data_vio->new_mapped.pbn,
+			(unsigned long long) data_vio->new_mapped.pbn,
 			thread_id,
 			expected);
 }
@@ -882,7 +882,7 @@ static inline void assert_data_vio_in_journal_zone(struct data_vio *data_vio)
 	thread_id_t thread_id = vdo_get_callback_thread_id();
 	ASSERT_LOG_ONLY((expected == thread_id),
 			"data_vio for logical block %llu on thread %u, should be on journal thread %u",
-			data_vio->logical.lbn,
+			(unsigned long long) data_vio->logical.lbn,
 			thread_id,
 			expected);
 }
@@ -928,7 +928,7 @@ static inline void assert_data_vio_in_packer_zone(struct data_vio *data_vio)
 	thread_id_t thread_id = vdo_get_callback_thread_id();
 	ASSERT_LOG_ONLY((expected == thread_id),
 			"data_vio for logical block %llu on thread %u, should be on packer thread %u",
-			data_vio->logical.lbn,
+			(unsigned long long) data_vio->logical.lbn,
 			thread_id,
 			expected);
 }
